Added quasi-one dimensional nozzle flow solver and postprocessing.
The solver will output the simulation information into a directory called "output."
It is suggested to run the solver inside the simulation folder so the
postprocessing codde can be used. The postprocessing code is inside the "plotting.py"
file inside the simulation folder. Just run plotting.py to create and visualize plots
from the simulation.